# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Health Diet Online Search",
      "description": "A community for Health experts and users"
    },
    "host": "localhost:3000",
    "basePath": "/api/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/posts/add": {
        "post": {
          "tags": ["Posts"],
          "summary": "add new post by expert",
          "description": "",
          "operationId": "addPost",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Post object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Post"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "post created successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/get/{id}": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get posts by id",
          "description": "",
          "operationId": "getPostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be fetched",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Post"
              }
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/get/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get total posts",
          "description": "",
          "operationId": "getPosts",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 posts per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Post"
              }
            },
            "404": {
              "description": "No posts found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/get/myposts/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Posts"],
          "summary": "Get total posts of the logined expert",
          "description": "",
          "operationId": "getPostsOfLogedinExpert",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 posts per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Post"
              }
            },
            "404": {
              "description": "No posts found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/update/{id}": {
        "put": {
          "tags": ["Posts"],
          "summary": "update post",
          "description": "",
          "operationId": "updatePostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Post object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Post"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This post doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/posts/delete/{id}": {
        "delete": {
          "tags": ["Posts"],
          "summary": "delete post by id",
          "description": "",
          "operationId": "deletePostById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of post to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "post deleted"
            },
            "401": {
              "description": "access denied."
            },
            "404": {
              "description": "The post with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
      "/feedbacks/add": {
        "post": {
          "tags": ["Feedbacks"],
          "summary": "add new feedback",
          "description": "",
          "operationId": "addFeedback",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "types are solution,question,post",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Feedback"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "feedback updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/feedbacks/update/{id}": {
        "put": {
          "tags": ["Feedbacks"],
          "summary": "update feedback",
          "description": "",
          "operationId": "updateFeedback",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "types are solution,question,post",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Feedback"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "feedback updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/feedbacks/delete/{id}": {
        "delete": {
          "tags": ["Feedbacks"],
          "summary": "delete feedback by id",
          "description": "",
          "operationId": "deleteFeedbackById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of feedback to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }

          ],
          "responses": {
            "200": {
              "description": "feedback deleted"
            },
            "404": {
              "description": "The feedback with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
...
        "Post": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "data": {
              "type": "string"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
            
          }
        },
...
        "Feedback": {
          "type": "object",
          "properties": {
            "feedback": {
              "type": "string"
            },
            "postId" : {
              "type" : "string"
            },
            "type" :{
              "type" : "solution | post"
            }
          }
        },
...
      } 
}